feat(f1r,f2): F1R evidence frozen + F2 paired BF16 clean canary
F1R Final:
- Timing fixed: 10 per-frame synchronized repeats, per-frame median/p90
- Eager: median=0.231s p90=0.287s VRAM=14.09GiB
- Flash2: median=0.198s p90=0.245s VRAM=14.09GiB
- Flash2 17% faster, identical VRAM
- Token match: 9/10 (only t00_s10_step0, pos=5/rz, margin=0.125)
- Gripper match: 10/10
- Determinism: 10/10 both profiles

Single divergence margin audit:
- pos=5(rz): Eager top1=31858(19.375) top2=31876(19.250) margin=0.125
- pos=5(rz): Flash2 top1=31876(19.250) top2=31858(19.125) margin=0.125
- NOT gripper dimension. Gripper margin=21.75(eager)/23.25(flash2)
- Harmless numerical boundary flip at near-tie argmax

F2 Paired BF16 Clean Canary (explicit center-crop):
  task0_init0: Eager=87 SUCCESS  Flash2=79 SUCCESS (diff steps)
  task0_init1: Eager=91 SUCCESS  Flash2=220 TIMEOUT (DIVERGENCE)
  task1_init0: Eager=109 SUCCESS Flash2=109 SUCCESS (EXACT MATCH)
  task1_init1: Eager=106 SUCCESS Flash2=99 SUCCESS (diff steps)

Classification: FLASH2_DISTINCT_CLOSED_LOOP_PROFILE (Case F2-C)
→ task0_init1 reveals Flash2 can diverge on sensitive init states
→ Cannot silently replace Eager with Flash2 for attacks
→ FP32 also fails on task0_init1 (this init state is multi-profile-sensitive)

+++ scripts/infra/run_spatial_closed_loop_canary.py
@@ -116,7 +116,7 @@ def run_episode(model, proc, task_suite, task_idx, init_idx, output_dir, label,
 
         inputs = proc(prompt, processed_img, return_tensors="pt")
         ids = inputs["input_ids"].to(device=DEV)
-        px = inputs["pixel_values"].to(dtype=torch.float32, device=DEV)
+        px = inputs["pixel_values"].to(dtype=dtype, device=DEV)
         pixel_sha = hashlib.sha256(px.float().cpu().numpy().tobytes()).hexdigest()
         ids_sha = hashlib.sha256(ids.cpu().numpy().tobytes()).hexdigest()
